% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{antipodes Server}
\usage{
server(session, input, output)
}
\arguments{
\item{session}{Shiny session object}

\item{input}{Shiny input object}

\item{output}{Shiny output object}
}
\description{
Simple server function for Shiny app
}
